Change default random seed to long.

+++ src/test/java/com/datumbox/configuration/TestConfiguration.java
@@ -30,7 +30,7 @@ public class TestConfiguration {
     public static final double DOUBLE_ACCURACY_MEDIUM = 0.01;
     public static final double DOUBLE_ACCURACY_LOW = 0.5;
     
-    public static final int RANDOM_SEED = 42;
+    public static final long RANDOM_SEED = 42L;
     
     public static final Class<? extends DatabaseConfiguration> PERMANENT_STORAGE = InMemoryConfiguration.class; 
     //public static final Class<? extends DatabaseConfiguration> PERMANENT_STORAGE = MapDBConfiguration.class;
